feat: seperate create modal
