+++ client/src/component/molecule/CreateModal/index.ts
@@ -1,0 +1,11 @@
+export { default } from './CreateModal'
+
+export interface CreateModalProps {
+  workspaceId: number
+  createModal: boolean
+  setCreateModal: Function
+}
+
+export interface ModalProps {
+  onSuccess?: () => void
+}
